Updated login functionality for trainer
