+++ yotacli/src/components/trainer/usermanagement/ForgotPassword.js
@@ -2,61 +2,42 @@ import Form from 'react-bootstrap/Form';
 import { Link, useNavigate } from 'react-router-dom';
 import loginCss from './Login.css'
 import LoginHeader from './LoginHeader';
-import { ToastContainer, toast } from 'react-toastify';
-import 'react-toastify/dist/ReactToastify.css';
-import { useState } from 'react';
 
-const Login = () => {
-    const navigate = useNavigate();
-    const [email, setEmail] = useState('');
-    const [password, setPassword] = useState('');
-
-    const loginUser = () => {
-        console.log("email: ", email)
-        let loginData = {
-            email: email,
-            password: password
-        }
-        console.log(loginData);
-        console.log("Login Data: " + loginData.email);
-        //Can call API here to verify the login details:
-        toast("Login Success!");
-        setTimeout(() => {
-            navigate('/trainer');
-        }, 1000);
-    }
-
+const ForgotPassword = () => {
     return (
         <>
             <LoginHeader />
-            <ToastContainer />
             <div className='login-bg-body'>
                 <div class="container">
                     <div class="row">
                         <div class="col"> </div>
                         <div class="card login-card-body">
                             <div class="card-header">
-                                <h3 className='text-center'>Trainer Login</h3>
+                                <h3 className='text-center'>Reset Password</h3>
                             </div>
                             <div class="card-body">
                                 <div class="col mt-4">
-                                    <Form>
+                                    <div>
+                                        <h3>In Progress...</h3>
+                                        <Link to='/login'>Back</Link>
+                                    </div>
+                                    {/* <Form>
                                         <Form.Group className="mb-3" controlId="formGroupEmail">
                                             <Form.Label>Email Address:</Form.Label>
-                                            <Form.Control type="email" onChange={(e)=>{setEmail(e.target.value)}} placeholder="Enter email" />
+                                            <Form.Control type="email" placeholder="Enter email" />
                                         </Form.Group>
                                         <Form.Group className="mb-3" controlId="formGroupPassword">
                                             <Form.Label>Password:</Form.Label>
-                                            <Form.Control type="password" onChange={(e)=>{setPassword(e.target.value)}} placeholder="Password" />
+                                            <Form.Control type="password" placeholder="Password" />
                                         </Form.Group>
                                     </Form>
                                     <div className='text-center'>
-                                        <button type='submit' className='btn btn-success login-btn' onClick={loginUser}>Login</button>
+                                        <button type='submit' className='btn btn-success login-btn' onClick={dashboard}>Login</button>
                                         <span className='d-flex justify-content-center'>
-                                            <Link to='/forgotPassword'>Forgot Password</Link>&nbsp;&nbsp;|
+                                            <Link to='#'>Forgot Password</Link>&nbsp;&nbsp;|
                                             New User?&nbsp;<Link to='/register'>Register</Link>
                                         </span>
-                                    </div>
+                                    </div> */}
                                 </div>
                             </div>
                         </div>
@@ -64,8 +45,7 @@ const Login = () => {
                     </div>
                 </div>
             </div>
-            {/* <Footer /> */}
         </>
     );
 }
-export default Login;
+export default ForgotPassword;
